requirements.txt
torch==2.4.0
torchvision==0.19.0
ultralytics==8.3.5
fastapi
uvicorn
Pillow
opencv-python-headless